moving gamestates to be on the same level as game.  They probably both need to go in game manager or or some other package.
