+++ gamemanager/gamemanager.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"solitaire/game"
-	"solitaire/game/gamestates"
+	"solitaire/gamestates"
 
 	"github.com/rs/xid"
 )
